fix(tool-search): expose internal batch executor
