+++ src/agents/tool-search.test.ts
@@ -28,6 +28,7 @@ import {
   resolveToolExecutionErrorKind,
 } from "./tool-result-error.js";
 import { compactToolSearchCatalogEntry } from "./tool-search-catalog.js";
+import { executeInternalToolSearchRequest } from "./tool-search-execute.js";
 import { readToolSearchRequest } from "./tool-search-request.js";
 import { ToolSearchRuntime } from "./tool-search-runtime.js";
 import {
@@ -419,7 +420,7 @@ describe("Tool Search", () => {
     const duplicateQueries = Array.from({ length: 16 }, () => ({
       query: "boundary duplicate",
     }));
-    const duplicateResult = await testing.executeInternalToolSearchRequest(
+    const duplicateResult = await executeInternalToolSearchRequest(
       { config, catalogRef },
       {
         queries: duplicateQueries,
@@ -428,7 +429,7 @@ describe("Tool Search", () => {
     expect(resultDetails(duplicateResult).results).toHaveLength(16);
     expect(catalogRef.current?.searchCount).toBe(16);
 
-    const clampedResult = await testing.executeInternalToolSearchRequest(
+    const clampedResult = await executeInternalToolSearchRequest(
       { config, catalogRef },
       {
         queries: Array.from({ length: 5 }, (_, index) => ({
@@ -455,7 +456,7 @@ describe("Tool Search", () => {
       catalogRef,
     });
     await expect(
-      testing.executeInternalToolSearchRequest(
+      executeInternalToolSearchRequest(
         { config, catalogRef },
         {
           queries: [{ query: "atomic validation" }, { query: " " }],
@@ -506,7 +507,7 @@ describe("Tool Search", () => {
       (candidate) => candidate.id,
     );
 
-    const result = await testing.executeInternalToolSearchRequest(
+    const result = await executeInternalToolSearchRequest(
       { config, catalogRef },
       {
         queries: Array.from({ length: 5 }, () => ({ query: "large surface", limit: 10 })),
@@ -528,7 +529,7 @@ describe("Tool Search", () => {
     }
 
     const manyGroups = resultDetails(
-      await testing.executeInternalToolSearchRequest(
+      await executeInternalToolSearchRequest(
         { config, catalogRef },
         {
           queries: Array.from({ length: 16 }, () => ({ query: "large surface", limit: 1 })),
@@ -565,7 +566,7 @@ describe("Tool Search", () => {
       catalogRef,
     });
     const result = resultDetails(
-      await testing.executeInternalToolSearchRequest(
+      await executeInternalToolSearchRequest(
         { config, catalogRef },
         {
           queries: Array.from({ length: 16 }, () => ({ query: "large remote surface", limit: 1 })),
@@ -607,7 +608,7 @@ describe("Tool Search", () => {
     const clientTool = fakeTool(`client_large_name_${"n".repeat(20_000)}`, "oversized metadata");
     addClientToolsToToolSearchCatalog({ tools: [clientTool], config, catalogRef });
     const result = resultDetails(
-      await testing.executeInternalToolSearchRequest(
+      await executeInternalToolSearchRequest(
         { config, catalogRef },
         {
           queries: Array.from({ length: 16 }, () => ({
@@ -681,7 +682,7 @@ describe("Tool Search", () => {
       expect.objectContaining({ name: "fake_attention", source: "openclaw" }),
     ]);
 
-    const batch = await testing.executeInternalToolSearchRequest(
+    const batch = await executeInternalToolSearchRequest(
       { config, catalogRef },
       {
         queries: [
@@ -722,7 +723,7 @@ describe("Tool Search", () => {
       config,
       catalogRef,
     });
-    const result = await testing.executeInternalToolSearchRequest(
+    const result = await executeInternalToolSearchRequest(
       { config, catalogRef },
       {
         queries: [{ query: "directory calendar", limit: 1 }],
